+++ module/entity/entity_manga/lib/entity_manga.dart
@@ -11,6 +11,7 @@ export 'src/enum/manga_menu_enum.dart';
 export 'src/enum/source_enum.dart';
 export 'src/exception/data_not_found_exception.dart';
 export 'src/exception/failed_parsing_html_exception.dart';
+export 'src/extension/source_search_manga_parameter_extension.dart';
 export 'src/extra/search_parameter_extra.dart';
 export 'src/history.dart';
 export 'src/manga.dart';
